[quant] QoL improvements for pipeline-level quant config (#11876)
* add repr for pipelinequantconfig.

* update
